Implemented test suite for UDP receiver
